BUG: Changed the type of Image::Origin and Spacing from float to double, and fixed up all the compilation errors that this caused elsewhere.  This change will provide much improved numerical precision at negligible cost in performance (because double is rarely faster than float on modern workstations and there are very few origin/spacing values compared to the number of pixel values).

+++ Testing/Code/Common/itkIteratorTests.cxx
@@ -29,8 +29,8 @@ void main()
   typedef itk::Image<itk::Scalar<unsigned short>, 3> ScalarImage;
   ScalarImage::Pointer  o3 = ScalarImage::New();
 
-  float origin3D[3] = { 5, 2.1, 8.1};
-  float spacing3D[3] = { 1.5, 2.1, 1};
+  double origin3D[3] = { 5, 2.1, 8.1};
+  double spacing3D[3] = { 1.5, 2.1, 1};
 
   ScalarImage::SizeType imageSize3D = { 200, 200, 200 };
   ScalarImage::SizeType bufferSize3D = { 200, 200, 200 };
